Refactor ListView into separate list and entry components
